refactor(user-location): refactor userlocation code into services
